Updates :- (#50)
1. Updated int64 to int.
2. Updated port to 4000

+++ internal/docgenerator/docgenerator.go
@@ -13,7 +13,7 @@ const (
 )
 
 const DefaultKeySize int = 250
-const DefaultDocSize int64 = 1024
+const DefaultDocSize int = 1024
 
 // Generator helps to generate random document for inserting and updating random
 // as per the doc loading task requirement.
@@ -21,12 +21,12 @@ type Generator struct {
 	DocType   string
 	KeyPrefix string
 	KeySuffix string
-	Seed      int64
-	SeedEnd   int64
+	Seed      int
+	SeedEnd   int
 	Template  template.Template
 }
 
-func ConfigGenerator(doctype, keyPrefix, keySuffix string, seed, seedEnd int64, template template.Template) *Generator {
+func ConfigGenerator(doctype, keyPrefix, keySuffix string, seed, seedEnd int, template template.Template) *Generator {
 	return &Generator{
 		DocType:   doctype,
 		KeyPrefix: keyPrefix,
@@ -38,12 +38,12 @@ func ConfigGenerator(doctype, keyPrefix, keySuffix string, seed, seedEnd int64, 
 }
 
 // GetDocIdAndKey will return key for the next document
-func (g *Generator) GetDocIdAndKey(iteration int64) (string, int64) {
+func (g *Generator) GetDocIdAndKey(iteration int) (string, int) {
 	newKey := iteration + g.SeedEnd
 	return fmt.Sprintf("%s%d%s", g.KeyPrefix, newKey, g.KeySuffix), newKey
 }
 
 // BuildKey returns the formatted key with unique identifier.
-func (g *Generator) BuildKey(key int64) string {
+func (g *Generator) BuildKey(key int) string {
 	return fmt.Sprintf("%s%d%s", g.KeyPrefix, key, g.KeySuffix)
 }
