+++ internal/tasks/task.go
@@ -3,7 +3,7 @@ package tasks
 type Task interface {
 	Describe() string
 	Do() error
-	Config(req *Request, seed int64, seedEnd int64, reRun bool) (int64, error)
+	Config(req *Request, seed int, seedEnd int, reRun bool) (int, error)
 	BuildIdentifier() string
 	CheckIfPending() bool
 	tearUp() error
