+++ internal/tasks/task_result.go
@@ -20,7 +20,7 @@ func (r *TaskResult) Do() error {
 	return nil
 }
 
-func (r *TaskResult) Config(req *Request, seed int64, seedEnd int64, rerun bool) (int64, error) {
+func (r *TaskResult) Config(req *Request, seed int, seedEnd int, rerun bool) (int, error) {
 	r.TaskPending = false
 	return 0, nil
 }
