+++ internal/tasks/task_single_upsert.go
@@ -22,7 +22,7 @@ type SingleUpsertTask struct {
 	InsertOptions   *InsertOptions          `json:"insertOptions,omitempty" doc:"true"`
 	OperationConfig *SingleOperationConfig  `json:"singleOperationConfig" doc:"true"`
 	Operation       string                  `json:"operation" doc:"false"`
-	ResultSeed      int64                   `json:"resultSeed" doc:"false"`
+	ResultSeed      int                     `json:"resultSeed" doc:"false"`
 	TaskPending     bool                    `json:"taskPending" doc:"false"`
 	result          *task_result.TaskResult `json:"-" doc:"false"`
 	req             *Request                `json:"-" doc:"false"`
@@ -55,7 +55,7 @@ func (task *SingleUpsertTask) CheckIfPending() bool {
 }
 
 // Config configures  the insert task
-func (task *SingleUpsertTask) Config(req *Request, seed int64, seedEnd int64, reRun bool) (int64, error) {
+func (task *SingleUpsertTask) Config(req *Request, seed int, seedEnd int, reRun bool) (int, error) {
 	task.TaskPending = true
 	task.req = req
 
@@ -71,7 +71,7 @@ func (task *SingleUpsertTask) Config(req *Request, seed int64, seedEnd int64, re
 	}
 
 	if !reRun {
-		task.ResultSeed = time.Now().UnixNano()
+		task.ResultSeed = int(time.Now().UnixNano())
 		task.Operation = SingleUpsertOperation
 		task.result = task_result.ConfigTaskResult(task.Operation, task.ResultSeed)
 
@@ -125,7 +125,7 @@ func (task *SingleUpsertTask) Do() error {
 
 	singleUpsertDocuments(task, collection)
 
-	task.result.Success = int64(len(task.OperationConfig.KeyValue)) - task.result.Failure
+	task.result.Success = (len(task.OperationConfig.KeyValue)) - task.result.Failure
 
 	if err := task.result.SaveResultIntoFile(); err != nil {
 		log.Println("not able to save result into ", task.ResultSeed)
